Remove Magic Extract
Magic extract was just meant as a placeholder until we built lock and extract, both of which are fully functioning. As such, we are should now remove magic extract so it doesn't continue to muddle up real production-ready code.

We also re-organize tests to make it easier to see the natural divisions.
